Change #1 - package naming
기존 프로젝트의 소스코드를 복사해 왔기 때문에, 패키지명을 변경
