+++ src/main/java/com/fastcampus/spring/logic/BubbleSort.java
@@ -1,8 +1,11 @@
-package com.fastcampus.java.logic;
+package com.fastcampus.spring.logic;
+
+import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.List;
 
+@Component
 public class BubbleSort <T extends Comparable<T>> implements Sort<T> {
 
     @Override
