+++ src/main/java/com/fastcampus/spring/logic/JavaSort.java
@@ -1,9 +1,12 @@
-package com.fastcampus.java.logic;
+package com.fastcampus.spring.logic;
+
+import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
+@Component
 public class JavaSort <T extends Comparable<T>> implements Sort<T> {
 
     @Override
