+++ src/test/java/com/fastcampus/spring/MainTest.java
@@ -1,4 +1,4 @@
-package com.fastcampus.java;
+package com.fastcampus.spring;
 
 import org.junit.jupiter.api.Test;
