+++ src/test/java/com/fastcampus/spring/logic/BubbleSortTest.java
@@ -1,4 +1,4 @@
-package com.fastcampus.java.logic;
+package com.fastcampus.spring.logic;
 
 import org.junit.jupiter.api.Test;
 
@@ -6,15 +6,15 @@ import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class JavaSortTest {
+class BubbleSortTest {
 
     @Test
-    void javaSortTest() {
+    public void bubbleSortTest() {
         // given
-        JavaSort<Integer> javaSort = new JavaSort<>();
+        BubbleSort<Integer> bubbleSort = new BubbleSort<>();
 
         // when
-        List<Integer> actual = javaSort.sort(List.of(1, 2, 4, 3, 5));
+        List<Integer> actual = bubbleSort.sort(List.of(1, 3, 4, 5, 2));
 
         // then
         assertEquals(List.of(1, 2, 3, 4, 5), actual);
